chore(web): remove graphqlQueryFn
