+++ web/src/utils/getGraphqlUrl.ts
@@ -1,0 +1,11 @@
+import { arbitrumSepolia } from "wagmi/chains";
+
+const CHAINID_TO_DISPUTE_TEMPLATE_SUBGRAPH = {
+  [arbitrumSepolia.id]:
+    process.env.REACT_APP_DRT_ARBSEPOLIA_SUBGRAPH ?? "Wrong Subgraph URL. Please check the environment variables.",
+};
+
+export const getGraphqlUrl = (isDisputeTemplate = false, chainId: number = arbitrumSepolia.id) => {
+  const coreUrl = process.env.REACT_APP_CORE_SUBGRAPH ?? "Wrong Subgraph URL. Please check the environment variables.";
+  return isDisputeTemplate ? CHAINID_TO_DISPUTE_TEMPLATE_SUBGRAPH[chainId] : coreUrl;
+};
